requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
jinja2>=3.1.0
httpx>=0.27.0
PyMuPDF>=1.24.0
Pillow>=10.0.0
pyyaml>=6.0.0
